+++ R/extend.R
@@ -210,7 +210,7 @@ extend_se.data.frame <- function(source, assignments,
 #' Create a node similar to a Codd extend relational operator (add derived columns).
 #'
 #' Partitionby and orderby can only be used with a database that supports window-functions
-#' (such as PostgreSQL, Spark, and so on).
+#' (such as PostgreSQL, Spark, and so on).  extend_nse() used bquote() .()-style abstraction.
 #'
 #' @param source source to select from.
 #' @param ... new column assignment expressions.
@@ -227,7 +227,9 @@ extend_se.data.frame <- function(source, assignments,
 #'   my_db <- DBI::dbConnect(RSQLite::SQLite(), ":memory:")
 #'   d <- rq_copy_to(my_db, 'd',
 #'                    data.frame(AUC = 0.6, R2 = 0.2))
-#'   optree <- extend_nse(d, v %:=% ifelse(AUC>0.5, R2, 1.0))
+#'   NEWCOL <- as.name("v")
+#'   NEWVALUE = "zz"
+#'   optree <- extend_nse(d, .(NEWCOL) %:=% ifelse(AUC>0.5, R2, 1.0), .(NEWVALUE) %:=% 6)
 #'   cat(format(optree))
 #'   sql <- to_sql(optree, my_db)
 #'   cat(sql)
@@ -261,6 +263,7 @@ extend_nse.relop <- function(source,
   # Recommend way to caputre ... unevalauted from
   # http://adv-r.had.co.nz/Computing-on-the-language.html#substitute "Capturing unevaluated ..."
   exprs <-  eval(substitute(alist(...)))
+  exprs <- lapply_bquote_to_langauge_list(exprs, env)
   if(length(setdiff(reverse, c(orderby, partitionby)))>0) {
     stop("rquery::extend_nse.relop all reverse columns must also be orderby or partitionby columns")
   }